requirements.txt
crewai>=0.50.0
crewai-tools>=0.4.0
langchain>=0.2.10
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
rich>=13.7.0
streamlit>=1.30.0
Pillow>=10.0.0
